+++ test/integration/tests/button/style.js
@@ -61,41 +61,48 @@ describe('paypal button color', () => {
     });
 
     it('should render a button with gold background when no color is specified', () => {
-        window.paypal.Buttons({
-
-            test: {
-                onRender() {
-                    assert.ok(getElementRecursive('.paypal-button-color-gold'));
-                }
+        return window.paypal.Buttons({
+            style: {
+                color: ''
             }
-
-        }).render('#testContainer');
+        }).render('#testContainer').then(() => {
+            assert.ok(getElementRecursive('.paypal-button-color-gold'));
+        });
     });
 
     it('should render a button with black background when passed "black"', () => {
-        window.paypal.Buttons({
-
-            test: {
-                style: {
-                    color: 'black'
-                },
-                onRender() {
-                    assert.ok(getElementRecursive('.paypal-button-color-black'));
-                }
+        return window.paypal.Buttons({
+            style: {
+                color: 'black'
             }
-        }).render('#testContainer');
+        }).render('#testContainer').then(() => {
+            assert.ok(getElementRecursive('.paypal-button-color-black'));
+        });
     });
 
-    it('should render a button with gold background when passed empty string', (done) => {
+    it('should render a button with gold background when passed ""', () => {
+        return window.paypal.Buttons({
+            style: {
+                color: ''
+            }
+        }).render('#testContainer').then(() => {
+            assert.ok(getElementRecursive('.paypal-button-color-gold'));
+        });
+    });
+
+    it('should not mutate the style object', (done) => {
+        const style = {
+            shape: 'pill'
+        };
+        const expected = JSON.stringify(style);
         done = once(done);
         window.paypal.Buttons({
-
+            style,
             test: {
-                style: {
-                    color: ''
-                },
                 onRender() {
-                    assert.ok(getElementRecursive('.paypal-button-color-gold'));
+                    if (JSON.stringify(style) !== expected) {
+                        done(new Error(`Expected style object ${ JSON.stringify(style) } to remain unmodified as ${ expected }`));
+                    }
                     done();
                 }
             },